XUnitSample - Updated

+++ CSharpSamples/XUnitSample/src/XUnitSample.Infrastructure/Services/Interfaces/ICalculatorService.cs
@@ -5,5 +5,5 @@ public interface ICalculatorService
     public int AddTwoNumbers(int firstNumber, int secondNumber);
     public int SubtractTwoNumbers(int firstNumber, int secondNumber);
     public int MultiplyTwoNumbers(int firstNumber, int secondNumber);
-    public int DivideTwoNumbers(int firstNumber, int secondNumber);
+    public double DivideTwoNumbers(double firstNumber, double secondNumber);
 }
